computer/requirements.txt
# Core dependencies
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-multipart>=0.0.6
aiosqlite>=0.19.0
aiofiles>=24.1.0
httpx>=0.26.0
pyyaml>=6.0.1
python-frontmatter>=1.1.0
sse-starlette>=2.0.0
claude-agent-sdk>=0.1.19
watchfiles>=0.21.0
watchdog>=4.0.0
apscheduler>=3.10.0
